% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_manipul_main.R
\name{all_char_paths}
\alias{all_char_paths}
\title{Summarizes path of all ancestors (names) for every annotate character}
\usage{
all_char_paths(ontology, sep = " | ")
}
\arguments{
\item{ontology}{ontology_index object with character annatotions included (ontology$annot_characters).}

\item{sep}{separator use to delimit ontology terms}
}
\value{
The matrix.
}
\description{
Returns matrix with character ids and their ancestor paths
}
\examples{
chr_paths=all_char_paths(ontology)
write.csv(chr_paths, file="full_char_annots_ontology.csv")
}
