% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontology_operations_main.R
\name{annot_all_chars}
\alias{annot_all_chars}
\title{Annotates all character statement against all ontology terms}
\usage{
annot_all_chars(ontology, use.synonyms = TRUE, min_set = TRUE)
}
\arguments{
\item{ontology}{ontology_index object with character names (ontology$name_characters) and ids (ontology$id_characters)}

\item{use.synonyms}{using synonyms list during search. It has to be included in ontology,
see syn_extract()}

\item{min_set}{if TRUE eliminates higher order inferred ontology terms}
}
\value{
The list of matched ontology terms and their charatcter ids.
}
\description{
Matches character statement and returns most similar ontology terms using
grep and distance-based matching
}
\examples{
grep_all_chars=annot_all_chars(hao.obo)

}
