% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontology_operations_main.R
\name{annot_char_grep}
\alias{annot_char_grep}
\title{Annotates character statement against all ontology terms useng grep}
\usage{
annot_char_grep(ontology, char.statement, use.synonyms = TRUE,
  min_set = TRUE)
}
\arguments{
\item{ontology}{ontology_index object.}

\item{char.statement}{character statement}

\item{use.synonyms}{using synonyms list during search. It has to be included in ontology,
see syn_extract()}

\item{min_set}{if TRUE eliminates higher order inferred ontology terms}
}
\value{
The vector of matches ontology terms.
}
\description{
Matches character statement and returns most similar ontology terms
}
\examples{
annot_char_grep(hao.obo, "Mola on right mandible")

}
