% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_manipul_main.R
\name{chars_per_term}
\alias{chars_per_term}
\title{Get number of chracters per each ontology term}
\usage{
chars_per_term(ontology)
}
\arguments{
\item{ontology}{ontology_index object with character annatotions included (ontology$annot_characters).}
}
\value{
The matrix of ontology terms IDs, their names and character number.
}
\description{
Returns matrix summarizing  number of characters per each ontology terms in descending order
}
\examples{
ontology<-HAO
ontology$annot_characters<-list(`CHAR:1`=c("HAO:0000653"), `CHAR:2`=c("HAO:0000653"))
chars_per_term(ontology)
}
