% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontology_operations_main.R
\name{comp.sets}
\alias{comp.sets}
\title{Compares  orignal annotationd vs. auto_annotations}
\usage{
comp.sets(set_org, set_annot)
}
\arguments{
\item{set_org}{manula annotations}

\item{set_annot}{auto-annotations}
}
\value{
List.
}
\description{
Gives set relatinships
}
\examples{
comp.sets(set_org, set_annot)

}
