% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontology_operations_main.R
\name{compare_annot}
\alias{compare_annot}
\title{Compares  orignal annotationd vs. auto_annotations and returns a vector of relatinships summarized by comp.sets()}
\usage{
compare_annot(ontology, annot.auto)
}
\arguments{
\item{ontology}{ontology with ontology$annot_characters}

\item{annot.auto}{annot.auto=grep_all_chars; from grep search}
}
\value{
Vector with relatinships; each name is a character ID.
}
\description{
Compares  orignal annotationd vs. auto_annotations and returns a vector of relatinships summarized by comp.sets().
}
\examples{
com_lists=compare_annot(ontology, grep_all_chars)
}
