% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_manipul_main.R
\name{edges2list}
\alias{edges2list}
\title{Converts edge matrix of charater annotations to list}
\usage{
edges2list(edge.matrix)
}
\arguments{
\item{edge.matrix}{Two-columns edge matrix.}
}
\value{
The list.
}
\description{
Takes two columns edge matrix (columns from and two) and produces list
}
\examples{
annot_list<-list(`CHAR:1`=c("HAO:0000933", "HAO:0000958"), `CHAR:2`=c("HAO:0000833", "HAO:0000258"))
edge.matrix<-list2edges(annot_list)
edges2list(edge.matrix)
}
