% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontology_operations_main.R
\docType{data}
\name{exclude_terms}
\alias{exclude_terms}
\title{Ontology terms to exclude for sunburst plot}
\format{List containing ontological terms.}
\usage{
exclude_terms
}
\description{
List of ontology terms from Hymenoptera Anatomy ontology that migh be thought of
as redundant and excluded from some operations.
}
\examples{
exclude_terms
}
\keyword{datasets}
