% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontology_operations_main.R
\name{get_comp_table}
\alias{get_comp_table}
\title{Compares  orignal annotationd vs. auto_annotations and returns a table of comparisons}
\usage{
get_comp_table(ontology, grep_all_chars)
}
\arguments{
\item{ontology}{ontology with ontology$annot_characters}

\item{grep_all_chars}{auto annotated chars from grep search}
}
\value{
Table.
}
\description{
Compares  orignal annotationd vs. auto_annotations and returns a table of comparisons. depends on compare_annot()
}
\examples{
tb_compar=get_comp_table(hao.obo, grep_all_chars)
}
