% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_manipul_main.R
\name{get_descendants_chars}
\alias{get_descendants_chars}
\title{Get all character IDs descending from specified term}
\usage{
get_descendants_chars(ontology, terms, ...)
}
\arguments{
\item{ontology}{ontology_index object.}

\item{terms}{IDs of ontology terms for which descendants are queried.}

\item{...}{other parameters for ontologyIndex::get_descendants() function}
}
\value{
The vector of character IDs.
}
\description{
Returns all characters located (associated) with a given ontology terms
}
\examples{
get_descendants_chars(hao_obo, "HAO:0000653")
}
