% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontology_operations_main.R
\name{get_onto_id}
\alias{get_onto_id}
\title{Revise}
\usage{
get_onto_id(vec_name, ontology, names = F)
}
\arguments{
\item{annotated.char.list}{Character list with ontology annotations.}

\item{col_order_inverse}{The default creates the first columns consisting if character IDs and the second columns consisting of ontology annatotaions.
The inverse order changes the columns order.}
}
\value{
Two-columns matrix.
}
\description{
Takes list of charater annotations amd creates an edge matrix comprising two columns: from and to.
}
\examples{
list2edges(annotated.char.list, col_order_inverse=F)
}
