% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_manipul_main.R
\name{get_part_anc}
\alias{get_part_anc}
\title{Makes dataframe of descndanrts to plot using visNetwork}
\usage{
get_part_anc(ontology, terms, is_a = c("is_a"), part_of = c("BFO:0000050"),
  color = c("red", "blue"), all_links = F, incl.top.anc = T,
  highliht_focus = T)
}
\arguments{
\item{ontology}{ontology_index object.}

\item{terms}{temr id for which descendants to be displayed}

\item{is_a}{id of how is_a relationships are coded in ontology.
To no included in output use NA.}

\item{part_of}{same as previous}

\item{color}{color for is_a and part_of relationships}

\item{all_links}{whether all links (is_a and part_of) which link descendants with other nodes must be included in the output.
Better not use as it makes the output messy.}

\item{incl.top.anc}{include the parents of terms}

\item{highliht_focus}{whether terms mus be highlited}
}
\value{
The list of dataframes.
}
\description{
Returns a list of two dataframes: nodes and edges
}
\examples{
dt=get_part_descen(hao.obo, get_onto_id("mouthparts", ontology) , is_a=c("is_a"), part_of=c("BFO:0000050"))
visNetwork(dt$nodes, dt$edges, width = "100\%", height = "100\%") \%>\%
visNodes(borderWidthSelected=4)\%>\%
visOptions(highlightNearest = TRUE)\%>\%
visLayout(randomSeed = 12)
}
