% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_manipul_main.R
\name{list2edges}
\alias{list2edges}
\title{Converts list of charater annotations to two-column edge matrix}
\usage{
list2edges(annotated.char.list, col_order_inverse = F)
}
\arguments{
\item{annotated.char.list}{Character list with ontology annotations.}

\item{col_order_inverse}{The default creates the first columns consisting if character IDs and the second columns consisting of ontology annatotaions.
The inverse order changes the columns order.}
}
\value{
Two-columns matrix.
}
\description{
Takes list of charater annotations amd creates an edge matrix comprising two columns: from and to.
}
\examples{
list2edges(annotated.char.list, col_order_inverse=F)
}
