% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{myfun}
\alias{myfun}
\title{Add together two numbers.}
\usage{
myfun(x)
}
\arguments{
\item{x}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{
add(1, 1)
add(10, 1)
}
