% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_manipul_main.R
\name{onto_process}
\alias{onto_process}
\title{Shortcut to process characters and ontology}
\usage{
onto_process(ontology, name_characters, do.annot = TRUE, ...)
}
\arguments{
\item{ontology}{Ontology}

\item{name_characters}{a vector of character names}

\item{do.annot}{specifiees if you need to run automatic annotations or not}

\item{...}{other arguments for annot_all_chars() function}
}
\value{
Ontology index object named
}
\description{
This is a shortcut function to make characters and ontology suitable for visualization using ontoFAST interactive tools.
}
\examples{
## automatically preprocess ontology
# onto<-onto_process(HAO, Sharkey_2011[,1])
## make shiny_in object
# shiny_in<<-make_shiny_in(onto)
## run interactively to show only 50 characters
# runOntoFast(nchar=50, show.chars=T)
}
