% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runOntoFast}
\alias{runOntoFast}
\title{Runh shiny}
\usage{
runOntoFast(is_a = c("is_a"), part_of = c("BFO:0000050"), nchar, ...)
}
\arguments{
\item{is_a}{is_a}

\item{part_of}{part_of}
}
\value{
runs app.
}
\description{
Runh shiny
}
\examples{
runOntoFast(nchar=10)
}
