% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontology_operations_main.R
\docType{data}
\name{Sharkey_2011}
\alias{Sharkey_2011}
\title{Hymenoptera character statements}
\format{A data table with 392 rows; each row is a character statement with character states:
\describe{
  \item{CHARACTER STATEMENTS:}{Charater statements}
  \item{STATES}{sates of the character}
  \item{STATES}{....}
}}
\usage{
Sharkey_2011
}
\description{
A table of characters and character states from
Hymenoptera character matrix (Sharkey et al., 2011: Cladistics). The table contains 392 character statements.
}
\examples{
chars<-Sharkey_2011
# read .csv file directly
#char_et_states<-read.csv(system.file("data_onto", "Sharkey_2011.csv", package = "ontoFAST"),
# header=T,  stringsAsFactors = F, na.strings = "")
}
\references{
Sharkey, M.J., et al. 2011. Phylogenetic relationships among superfamilies of Hymenoptera. Cladistics 28(1), 80-112.
(\href{http://onlinelibrary.wiley.com/doi/10.1111/j.1096-0031.2011.00366.x/full}{Read})
}
\keyword{datasets}
