% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontology_operations_main.R
\docType{data}
\name{Sharkey_2011_annot}
\alias{Sharkey_2011_annot}
\title{Characters annotated with ontology terms}
\format{List containing characters and ontology terms.}
\usage{
Sharkey_2011_annot
}
\description{
List of character IDs and their ontology annotations. The characters are from Hymenoptera phylogeny (Sharkey et al., 2011: Cladistics).
The ontology annotations are from Hymenoptera Anatomy Ontology (HAO).
}
\examples{
Sharkey_2011_annot
}
\references{
Sharkey, M.J., et al. 2011. Phylogenetic relationships among superfamilies of Hymenoptera. Cladistics 28(1), 80-112.
(\href{http://onlinelibrary.wiley.com/doi/10.1111/j.1096-0031.2011.00366.x/full}{Read})
}
\keyword{datasets}
