% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontology_operations_main.R
\name{syn_extract}
\alias{syn_extract}
\title{Associates synonym names with ontology terms}
\usage{
syn_extract(ontology, list_id = "synonym")
}
\arguments{
\item{ontology}{ontology_index object.}

\item{list_id}{theu ID of list element where synonyms are stored}
}
\value{
The vector with ontology IDs and synonym names.
}
\description{
Extracts and parses synonyms from ontology to make them readable
}
\examples{
hao.obo$parsed_synonyms=syn_extract(hao.obo)

}
