% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontology_operations_main.R
\docType{data}
\name{Tarasov_2017_annot}
\alias{Tarasov_2017_annot}
\title{Dung beetle characters annotated with ontology terms}
\format{List containing characters and ontology terms.}
\usage{
Tarasov_2017_annot
}
\description{
List of character IDs and their ontology annotations. The characters are from Scarabaeinae phylogeny (Tarasov, 2017: Zootaxa).
The ontology annotations are from a modified version Hymenoptera Anatomy Ontology (HAO) that was enriched with additional terms (SCR:) to accommodate anatomy of the dung beetles.
}
\examples{
Tarasov_2017_annot
}
\references{
Tarasov, S., 2017. A cybertaxonomic revision of the new dung beetle tribe Parachoriini (Coleoptera: Scarabaeidae: Scarabaeinae) and its phylogenetic assessment using molecular and morphological data. Zootaxa, 4329(2), pp.101-149.
(\href{https://www.mapress.com/j/zt/article/view/zootaxa.4329.2.1}{Read})
}
\keyword{datasets}
