% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_manipul_main.R
\name{all_char_paths}
\alias{all_char_paths}
\title{Get all ontology ancestors for a character}
\usage{
all_char_paths(ontology, sep = " | ")
}
\arguments{
\item{ontology}{ontology_index object with character annatotions included (ontology$annot_characters).}

\item{sep}{separator use to delimit ontology terms}
}
\value{
The matrix.
}
\description{
Returns matrix with character ids and their ancestor paths
}
\examples{
#getting ontology
ontology<-HAO
# reading in characters
char_et_states<-Sharkey_2011
# embedding characters and character ids into ontology
id_characters<-paste("CHAR:",c(1:392), sep="")
name_characters<-char_et_states[,1]
names(name_characters)<-id_characters
ontology$name_characters<-name_characters
ontology$id_characters<-id_characters
ontology$annot_characters<-list(`CHAR:1`=c("HAO:0000653"), `CHAR:2`=c("HAO:0000653"))
#' all_char_paths(ontology)
}
