% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontology_operations_main.R
\name{annot_char_grep}
\alias{annot_char_grep}
\title{Annotate a character statement with ontology terms}
\usage{
annot_char_grep(ontology, char.statement, use.synonyms = TRUE, min_set = TRUE)
}
\arguments{
\item{ontology}{ontology.}

\item{char.statement}{character statement}

\item{use.synonyms}{if TRUE then the synonyms are used during search. The synonyms have to be included in the ontology using
syn_extract() function}

\item{min_set}{if TRUE eliminates higher order inferred ontology terms}
}
\value{
The vector of matches ontology terms.
}
\description{
Matches character statement and returns most similar ontology terms
}
\examples{
\dontrun{
annot_char_grep(HAO, "Mola on right mandible")
}
}
