% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_manipul_main.R
\name{chars_per_term}
\alias{chars_per_term}
\title{Get number of chracters per each ontology term}
\usage{
chars_per_term(ontology, annotations = "auto")
}
\arguments{
\item{ontology}{ontology_index object with character annatotions included (ontology$annot_characters).}

\item{annotations}{which annotations to use: "auto" means automatic annotations, "manual" means manual ones.
Alternatively, any othe list element containing annotations can be specified.}
}
\value{
The matrix of ontology terms IDs, their names and character number.
}
\description{
Returns matrix summarizing  number of characters per each ontology terms in descending order
}
\examples{
\dontrun{
ontology<-HAO
ontology$terms_selected_id<-list(`CHAR:1`=c("HAO:0000653"), `CHAR:2`=c("HAO:0000653"))
chars_per_term(ontology, annotations="manual")
}
}
