% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_manipul_main.R
\name{export_cytoscape}
\alias{export_cytoscape}
\title{Export to Cytoscape format}
\usage{
export_cytoscape(
  ontology,
  annotations = "auto",
  is_a = c("is_a"),
  part_of = c("BFO:0000050")
)
}
\arguments{
\item{ontology}{Ontology}

\item{annotations}{which annotations to use: "auto" means automatic annotations, "manual" means manual ones.
Alternatively, any othe list element containing annotations can be specified.}

\item{is_a}{is_a}

\item{part_of}{part_of}
}
\value{
Returns a table
}
\description{
This function converts character annotations to Cytoscape format. It returns a table that can be saved as in csv format
and imported using Cytoscape. In Cytoscape choose File -> Import -> Network -> File. The assign columns to nodes and edges. Do not select
columns that enumerated the tables' rows.
}
\examples{
\dontrun{
cyto<-export_cytoscape(shiny_in)
write.csv(cyto, "cyto_exp.csv")
}
}
