% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_manipul_main.R
\name{get_ancestors_chars}
\alias{get_ancestors_chars}
\title{Get all ancestal ontology terms in a given set of characters}
\usage{
get_ancestors_chars(ontology, char_id)
}
\arguments{
\item{ontology}{ontology_index object with character annatotions included (ontology$annot_characters).}

\item{char_id}{IDs of character.}
}
\value{
The vector of ontology terms IDs.
}
\description{
Returns all ontology terms which are ancestors of a given character set
}
\examples{
ontology<-HAO
ontology$annot_characters<-list(`CHAR:1`=c("HAO:0000653"), `CHAR:2`=c("HAO:0000653"))
get_ancestors_chars(ontology, c("CHAR:1","CHAR:2"))
}
