% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontology_operations_main.R
\name{get_onto_id}
\alias{get_onto_id}
\title{IDs for ontology term names}
\usage{
get_onto_id(vec_name, ontology, names = F)
}
\arguments{
\item{vec_name}{term names}

\item{names}{add element names}

\item{onto}{ontology}
}
\value{
vector.
}
\description{
Returns IDs of ontology terms given terms' names
}
\examples{
vec_name=c("ventral mesofurco-profurcal muscle", "anatomical entity")
get_onto_id(vec_name, HAO)
}
