% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontology_operations_main.R
\name{get_onto_name}
\alias{get_onto_name}
\title{Returns names of ontology terms for ontology IDs}
\usage{
get_onto_name(vec, onto, names = F)
}
\arguments{
\item{vec}{ID or IDs}

\item{onto}{ontology}

\item{names}{add element names}
}
\value{
vector.
}
\description{
Returns names of ontology terms for ontology IDs
}
\examples{
get_onto_name("HAO:0002272", hao.obo)
}
