% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ontology_operations_main.R
\name{get_onto_name}
\alias{get_onto_name}
\title{Get names for ontology IDs}
\usage{
get_onto_name(vec, onto, names = F)
}
\arguments{
\item{vec}{ID or a vector of IDs}

\item{onto}{ontology}

\item{names}{use element name}
}
\value{
vector of names.
}
\description{
Returns names of ontology terms for ontology IDs
}
\examples{
\dontrun{
get_onto_name("HAO:0002272", HAO)
}
}
