% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_manipul_main.R
\name{make_shiny_in}
\alias{make_shiny_in}
\title{Make ontology object for vizualization}
\usage{
make_shiny_in(ontology)
}
\arguments{
\item{ontology}{Ontology}
}
\value{
Ontology index object named as shiny_in.
}
\description{
Make ontology object for vizualization
}
\examples{
make_shiny_in(hao.obo)
}
