% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_manipul_main.R
\name{onto_process}
\alias{onto_process}
\title{Shortcut to process characters and ontology}
\usage{
onto_process(ontology, name_characters, do.annot = TRUE, ...)
}
\arguments{
\item{ontology}{Ontology}

\item{name_characters}{a vector of character names}

\item{do.annot}{specifiees if you need to run automatic annotations or not}

\item{...}{other arguments for annot_all_chars() function}
}
\value{
Ontology index object named
}
\description{
This is a shortcut function to make characters and ontology suitable for visualization using ontoFAST interactive tools.
}
\examples{
data(Sharkey_2011)
data(HAO)
hao_obo<-onto_process(HAO, Sharkey_2011[,1], do.annot = FALSE)
ontofast <- new.env(parent = emptyenv())
ontofast$shiny_in <- make_shiny_in(hao_obo)
runOntoFast(is_a = c("is_a"), part_of = c("BFO:0000050"), shiny_in="shiny_in" )
}
