% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runOntoFast}
\alias{runOntoFast}
\title{Run OntoFAST interactively}
\usage{
runOntoFast(is_a = c("is_a"), part_of = c("BFO:0000050"), nchar = "all",
  show.chars = T, ...)
}
\arguments{
\item{is_a}{is_a}

\item{part_of}{part_of}

\item{nchar}{number of characters to show}

\item{show.chars}{Shows character statements}
}
\value{
runs app.
}
\description{
Runh shiny
}
\examples{
#shiny_in<-make_shiny_in(HAO)
#runOntoFast(show.chars=F)
}
